+++ R/plot_safety_of_advice.R
@@ -7,14 +7,14 @@
 #' @return A ggplot object visualizing the safety of advice for either one or multiple symptom checkers
 #' @examples
 #' data(symptomcheckRdata)
-#' get_safety_of_advice <- get_safety_of_advice(
+#' safety_of_advice <- get_safety_of_advice(
 #'   data = symptomcheckRdata,
 #'   triagelevel_correct = "Goldstandard_solution",
 #'   triagelevel_advice = "Triage_advice_from_app",
 #'   order_triagelevel = c("Emergency", "Non-Emergency", "Self-care"),
 #'   apps = "App_name"
 #'   )
-#' safety_plot <- plot_safety_of_advice(get_safety_of_advice)
+#' safety_plot <- plot_safety_of_advice(safety_of_advice)
 #'
 #' @export
 #' @import dplyr