+++ NEWS.md
@@ -1,8 +1,12 @@
 # symptomcheckR 0.1.1
 
-* Revised CRAN submission.
+-   The package now includes confidence intervals for all metrics. These can be obtained by setting the paramter CI == TRUE.
+-   The package now contains get_irr() to calculate the inter-rater reliability if multiple raters were involved in the evaluation.
 
+# symptomcheckR 0.1.1
+
+-   Revised CRAN submission.
 
 # symptomcheckR 0.1.0
 
-* Initial CRAN submission.
+-   Initial CRAN submission.